filter_rewrite: Moved `Teleserver` to `.core`.
